used env & add logger & relocate the file tree & add develop docker
1. Use env to choose the program execute status
2. Add looger module
3. relocate the files
4. add development of docker image
