+++ .env.default
@@ -1,2 +1,4 @@
-version=0.1.0
-discordToken=This is fake token
+ENV=develop
+VERSION=0.1.0
+DISCORDTOKEN=This is fake token
+STORAGEFOLDER="static"
