+++ .env.example
@@ -1,2 +1,4 @@
-version=the project version
-discordToken=The discord token
+ENV=The host environment (develop,production,withoutBot)
+VERSION=the project version
+DISCORDTOKEN=The discord token
+STORAGEFOLDER=storage the static files
