+++ protobuf/People.proto
@@ -1,0 +1,10 @@
+syntax="proto3";
+
+message Person{
+    int32 id=1;
+    string name=2;
+}
+
+message People{
+    repeated Person people=1;
+}
