+++ src/deps.ts
@@ -1,6 +1,6 @@
 export {
-  type ButtonComponent,
   type Bot,
+  type ButtonComponent,
   ButtonStyles,
   ChannelTypes,
   createBot,
@@ -9,5 +9,6 @@ export {
   MessageComponentTypes,
   startBot,
 } from "https://deno.land/x/discordeno@18.0.1/mod.ts";
-export { join } from "https://deno.land/std@0.224.0/path/mod.ts";
-// export { load } from "https://deno.land/std@0.224.0/dotenv/mod.ts";
+export { join } from "jsr:@std/path";
+export { load } from "jsr:@std/dotenv";
+export { existsSync } from "jsr:@std/fs";
