+++ src/module/feature/Dice.ts
